+++ southbridge/amd/amd8111/ac97.c
@@ -28,7 +28,7 @@
 #include <config.h>
 #include "amd8111.h"
 
-static void lpci_set_subsystem(struct device * dev, unsigned vendor, unsigned device)
+static void lpci_set_subsystem(struct device * dev, u16 vendor, u16 device)
 {
 	pci_write_config32(dev, PCI_SUBSYSTEM_VENDOR_ID, 
 		((device & 0xffff) << 16) | (vendor & 0xffff));
