+++ southbridge/amd/amd8111/acpi.c
@@ -208,7 +208,7 @@ static void acpi_enable_resources(struct device * dev)
 	
 }
 
-static void lpci_set_subsystem(struct device * dev, unsigned vendor, unsigned device)
+static void lpci_set_subsystem(struct device * dev, u16 vendor, u16 device)
 {
 	pci_write_config32(dev, 0x7c, 
 		((device & 0xffff) << 16) | (vendor & 0xffff));
