+++ southbridge/amd/amd8111/usb.c
@@ -34,7 +34,7 @@
 #include "amd8111.h"
 
 
-static void lpci_set_subsystem(struct device * dev, unsigned vendor, unsigned device)
+static void lpci_set_subsystem(struct device * dev, u16 vendor, u16 device)
 {
 	pci_write_config32(dev, 0x70, 
 		((device & 0xffff) << 16) | (vendor & 0xffff));
